Realtime: make example customizable: (#1332)
Trying to make it easy to customize/use your own agents and reuse the
UI.
